+++ examples/AN01034_using_the_device_control_library_over_usb/host/src/resource.h
@@ -1,8 +1,8 @@
-// Copyright 2016-2021 XMOS LIMITED.
+// Copyright 2016-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
-#ifndef __resource_h__
-#define __resource_h__
+#ifndef RESOURCE_H_
+#define RESOURCE_H_
 
 #define RESOURCE_ID 0x12
 
-#endif // __resource_h__
+#endif // RESOURCE_H_