+++ tests/xscope_device/src/support.h
@@ -1,7 +1,7 @@
-// Copyright 2016-2021 XMOS LIMITED.
+// Copyright 2016-2024 XMOS LIMITED.
 // This Software is subject to the terms of the XMOS Public Licence: Version 1.
-#ifndef __support_h__
-#define __support_h__
+#ifndef SUPPORT_H_
+#define SUPPORT_H_
 
 #include "control.h"
 
@@ -28,4 +28,4 @@ int check(const struct options &o,
 
 void drive_user_task_registration(chanend c_user_task[n], unsigned n);
 
-#endif // __support_h__
+#endif // SUPPORT_H_